92 dev styles problems in pipeline (#95)
Partially refactored our approach to inserting templates and stylesheets - now with each update we'll insert all stylesheet information from scratch, instead of trying to depend on stuff that is already in the database.
We now depend more on _properties.json and _stylesheets.json files in each stylesheets directory.
_stylesheets.json is used on the server, for dev purposes script automatically creates _stylesheets.dev.json file.
It also allowed us to have better naming for stylesheets folder - it no longer is "theme1" or "theme5", instead we just have "masterStyle" and "crystallicumDark"