73 dev fix urls for codespaces (#98)
devcontainer was defined, which sets up the codespaces. Now they can be used and they will work (initialize script automatically replaces localhost with whatever site name is generated for codespaces.
They are not super fast to initialize, but they behave the same way as our local environments, so we we'll be able to use them for i.e. quick checks for review or something like that.
Or you know - even actual development if for some reason machine we're usually developing on is not accessible.